No identifer > any idenfier in semantic version
If a project as beta and final releases of a version, `punic` used to choose the latest beta version over the final release.
This fixes this by prefering a empty identifer array over any identifier.
An example of this problem is the [XCGLogger](https://github.com/DaveWoodCom/XCGLogger) project with is 4.0 release.