Use workspace aswell
`Punic` now also works with .xcworkspace files.
It looks for them in the Checkout folder and can run `xcodebuild` with the `-workspace` command.
Because each workspace also has a project I need to avoid duplicate building of the same scheme by avoiding duplictate scheme names.
Workspaces are prefered over a project.
If I did understand the `carthage` code correctly they do it the same way.
This change is necesairy to build `realm-cocoa`.
But you still need to add a skip for `RealmExamples`.